🚚 moved configuration files to config folder
